% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GLPK-constants}
\alias{GLPK-constants}
\alias{GLP_FR}
\alias{GLP_LO}
\alias{GLP_UP}
\alias{GLP_DB}
\alias{GLP_FX}
\alias{GLP_MIN}
\alias{GLP_MAX}
\title{GLPK constants}
\format{An object of class \code{integer} of length 1.}
\usage{
GLP_FR

GLP_LO

GLP_UP

GLP_DB

GLP_FX

GLP_MIN

GLP_MAX
}
\description{
GLPK constants
}
\section{Boundaries}{

\itemize{
\item \code{GLP_FR}: 1, free (unbounded) variable.
\item \code{GLP_LO}: 2, variable with lower bound.
\item \code{GLP_UP}: 3, variable with upper bound.
\item \code{GLP_DB}: 4, double-bounded variable.
\item \code{GLP_FX}: 5, fixed variable.
}
}

\section{Type of problem}{

\itemize{
\item \code{GLP_MIN}: 1, Minimization
\item \code{GLP_MAX}: 2, Maximization
}
}

\keyword{datasets}
