% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glpk_example}
\alias{glpk_example}
\title{A mininmal example}
\usage{
glpk_example(obj, subj_lhs, subj_rhs, pname = "sample")
}
\arguments{
\item{obj}{Numeric vector of size \eqn{K}. Coeficients in the objective function.}

\item{subj_lhs}{Numeric matrix of size \eqn{K\times m}. Constraints.}

\item{pname}{Character scalar. Name of the LP.}
}
\description{
A mininmal example
}
\examples{
obj      <- c(10, 6, 4)
subj_lhs <- matrix(c(1, 10, 2, 1, 4, 2, 1, 5, 6), ncol = 3)
subj_rhs <- c(100, 600, 300)
glpk_example(obj, subj_lhs, subj_rhs)
}
